Delete API v1 package
